@@ -30,53 +30,53 @@ dependencies:
   image_picker: ^1.0.2
 
   # state
-  hooks_riverpod: ^2.4.9
-  flutter_hooks: ^0.20.4
+  hooks_riverpod: ^2.5.1
+  flutter_hooks: ^0.20.5
   freezed_annotation: ^2.4.1
-  riverpod_annotation: ^2.0.0
-  json_annotation: ^4.8.1
-  shared_preferences: ^2.2.2
-  bot_toast: ^4.1.1
-  skeletonizer: ^0.8.0
+  riverpod_annotation: ^2.3.5
+  json_annotation: ^4.9.0
+  shared_preferences: ^2.2.3
+  bot_toast: ^4.1.3
+  skeletonizer: ^1.3.0
 
   # util
-  collection: ^1.17.2
-  timeago: ^3.5.0
+  collection: ^1.18.0
+  timeago: ^3.6.1
   faker: ^2.1.0
   kosher_dart: ^2.0.16
 
   # routing
-  go_router: ^10.1.2
+  go_router: ^14.1.4
 
   # networking
   dio: ^5.4.3+1
-  requests_inspector: ^4.0.0
+  requests_inspector: ^4.0.3
   http: ^1.1.0 # todo: remove
 
   # integrations
   rxdart: ^0.27.7
-  flutter_svg: ^2.0.7
+  flutter_svg: ^2.0.10+1
   cached_network_image: ^3.3.1
   widget_to_marker: ^1.0.5
   timezone: ^0.9.2
   flutter_timezone: ^1.0.7
 
   # icons
-  fluentui_system_icons: ^1.1.234
+  fluentui_system_icons: ^1.1.241
 
   # widgets
   flutter_easyloading: ^3.0.5 # todo: remove
-  syncfusion_flutter_charts: ^24.1.41
-  syncfusion_flutter_gauges: ^24.1.41
+  syncfusion_flutter_charts: ^26.1.35
+  syncfusion_flutter_gauges: ^26.1.35
   pin_code_fields: ^8.0.1
   dropdown_button2: ^2.3.9
   dotted_border: ^2.1.0
-  extended_text: ^13.0.0
-  
+  extended_text: ^13.0.2 # only used in 1 place (report_details_screen.dart)
+
   # cicd
   logger: ^2.3.0
-  sentry_flutter: ^7.14.0
-  device_info_plus: ^9.1.2
+  sentry_flutter: ^8.2.0
+  device_info_plus: ^10.1.0
   flutter_background_service: ^5.0.5
   flutter_background_service_android: ^6.2.2
   flutter_native_splash: ^2.4.0
@@ -88,25 +88,25 @@ dev_dependencies:
     sdk: flutter
 
   # analysis
-  flutter_lints: ^2.0.3
-  analyzer: ^6.2.0
-  custom_lint: ^0.5.11
-  riverpod_lint: ^2.3.7
+  flutter_lints: ^4.0.0
+  analyzer: 6.4.1
+  custom_lint: ^0.6.4
+  riverpod_lint: ^2.3.10
 
   # codegen
-  build_runner: ^2.4.6
-  freezed: ^2.4.2
-  riverpod_generator: ^2.3.9
-  flutter_gen_runner: ^5.3.1
-  json_serializable: ^6.7.1
-  go_router_builder: ^2.3.2
+  build_runner: ^2.4.11
+  freezed: ^2.5.2
+  riverpod_generator: ^2.4.0
+  flutter_gen_runner: ^5.5.0+1
+  json_serializable: ^6.8.0
+  go_router_builder: ^2.7.0
   
 dependency_overrides:
-  flutter_gen_core:
-    git:
-      url: https://github.com/FlutterGen/flutter_gen.git
-      ref: 8c57b5aa73f7793d614f064ec8e00434cc4cf01a
-      path: packages/core
+  # flutter_gen_core:
+  #   git:
+  #     url: https://github.com/FlutterGen/flutter_gen.git
+  #     ref: 8c57b5aa73f7793d614f064ec8e00434cc4cf01a
+  #     path: packages/core
   kosher_dart:
     git:
       url: https://github.com/noga-dev/kosher_dart.git
